Exporting: Enable field filters in scripts
We allow fields to be described and recorded within scripts. However, we
do not allow filtering on the fields. With the newly added field filter
closure support, we can now support filtering as well.

Add should_filter_field() and should_filter_record() methods to the
TimelineEventFlags struct. This allows either the local field values of
an event or the already recorded values of a timeline to be filtered.

When events are filtered in the timeline, prevent them from causing the
next action. For example, if the event should start a timeline, it will
only do so if the filters pass. If the event should end a timeline, it
will only do so if the filters pass. However, the timeline will be
removed even if the filter does not pass. This helps prevent timelines
that will never complete and cause excessive memory usage.

Update error statements to include the Event name to help debug longer
scripts that may get run with bad settings.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>