Scripting: Add OS version support
Often times scripts need OS version details, such as which OS they are
on (Windows, Linux, etc.) and what version (10.0, 6.11, etc.).

Add ScriptEnvironment struct that allows fetching the scripting engine
version and the OS version. Also move is_os() from it's own method to
this environment to make it clear this is the object to use going
forward.

Add new_environment() method for scripts that returns the new
ScriptEnvironment struct.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>