Exporting: Original Event Proxy Support
Often times scripts/tools want to record an original event and not have
it change at all. When this is done directly, such as via tracefs or
ETW, it's easy and we don't need any changes. However, when we proxy the
event from a sub-carrier event, like DotNet IPC, we need to be able to
describe events that have both a proxy_id as well as the original id.

Add flag to indicate record data is directly from an event. This allows
formats aware of this to not include sample values, etc. and just the
payload data.

When record_formats are created from an event, mark them with the
original data flag.

Update our proxy event pattern to require both a name and id for the
proxy event. Add optional proxy_id to Event struct and accessors.

Utilize these new flags and proxy support within DotNet IPC.

Move Guid to be a cross platform struct to allow DotNet to describe
things using that struct since it is multi-platform and now used in both
ETW parts and DotNet.

Persist provider Guid within Event via DotNet by adding the Guid to the
provider name. IE: ProviderName:{Guid}. This allows formats that wish to
use both the provider name and the Guid to parse them out without
causing every event to have a Guid stored within it (16 bytes).

Add SelfTest for cross platform Guid / provider name generation within
DotNet.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>