Exporting: Add Attributes Concept
Often times we want to sample data and also describe extra data on
per-event or per-record basis. For example, if we want to record an
ActivityId or OpCode from ETW, we have no place for this.

Introduce ExportAttributes struct (and sub structs) that allow
describing a set of attributes efficiently. All strings are interned and
stored as an ID. Attributes include a name and value, value can be a
pointer to a string by ID (Label) or an actual numeric value (Value).

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>