Exporting: Integrate attributes with trace context
Events added to the ExportMachine need to be able to create attributes
and link them up to samples that get created. DotNet requires proxy
support, and even if we proxy data out, the attributes need to be
associated with samples created in the separate proxy event.

Add push_association() to ExportAttributes struct. This allows events to
build up attributes that have been previously seen or to ensure proxy
events can still make their own attributes, but get them linked
together. This allows the pattern of common attributes to be saved once,
and only the unique parts of attributes be saved and also include common
attributes. When attributes with associations are fetched from the
ExportMachine, the caller receives all the pairs of attributes from that
attribute as well as all the associated attributes.

Add label_attribute() and value_attribute() to create attributes pairs
via the ExportTraceContext. This allows event callbacks to create
attribute pairs to be pushed into ExportAttributes struct.

Add push_unique_attributes() to assign an ExportAttributes struct to the
ExportMachine. The attribute ID given back can then be used with an
ExportProcessSample.

Add with_attributes() to ExportSampleBuilder struct. This allows event
callbacks to assign attributes they created to samples within the
ExportMachine.

Add ExportAttributeWalker stuct and use it within
ExportMachine::sample_attributes() to return attribute pairs. This
allows our ExportMachine to efficiently get all attributes, including
all associated attributes. This prevents cycles (if any) from causing
issues and ensures callers can re-use temporary Vec and HashSet for
optimal performance.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>