user_events: Fix mut compiler issue
User_events are created via an ABI that utilizes ioctl(). The libc crate
does not require the values to be passed as mutable, however, in the
case of the user_events ABI they do get mutated.

We did not mark the register struct as &mut and the compiler was allowed
to optimize the register struct write_index read to BEFORE the ioctl()
invocation. This caused the write_index to stay set to "unregistered" on
the Rust side. When the event was dropped, it would not run the
unregister path.

This lead to conditions that if Rust allocated another user_event Rust
struct at the same memory location (which we saw) the register would
start to fail.

Pass the user_event reg struct as &mut instead of & to avoid this.

Update user_event self test to a reasonable max event limit to ensure it
will pass on most systems that limit open file descriptor counts.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>