DotNet: Add set_dotnet_filter_args() script method
Scripts need to be able to set DotNet filter arguments on a
per-provider basis. When providers are enabled on Windows and Linux,
this value is passed and allows extra configuration.

For example, the System.Runtime EventSource allows the counter rate to
be set via EventCounterIntervalSec, such as "EventCounterIntervalSec=1".

Add set_dotnet_filter_args() to Rhai script engine to allow for this.

Add filter_args for Linux UserEvent based tracepoint integration.

Add stubs for filter_args on Windows, later commit will enable this.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>