ETW: Add Session ID to SessionCallbackContext
ETW has a session ID concept within the WNODE struct (HistoricalContext)
that can be used for unique identification of the session.

Some providers, like the CLR/DotNet, allow registry changes on a
per-session ID basis to set configurations. In order to do this, the
session ID is used. We need to ensure this can be read easily by ETW
session callbacks.

Add session ID via id() on the SessionCallbackContext. This allows the
various ETW state session closures (starting, started, stopping) to get
the session ID when needed. The session ID is no longer available (and
will default to 0) during the stopped closures.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>