Exporting: Allow OS native fields for Timeline IDs
Some attributes of events come directly from the OS native session. For
example, the process ID and the thread ID come via ETW structures
per-event callback. On Linux some of these are in the "common_fields",
such as "common_pid".

In order to allow scripts to treat this common between OS's and to allow
expansion in the future (such as the trace-context work in
OpenTelemetry) setup the pattern for virtual fields.

Add VirtualEventFields struct that encapsulates the logic on how to get
OS data from the TraceContext of an Event closure and copy it into a
slice.

Use VirtualEventFields struct in timeline field ID parsing to allow for
these to be used as IDs. Scripts can now use "os:pid" and "os:tid" as
field names for ID data across each OS. Filtering on these fields will
be added in future commits.

Add self test for these new virtual fields (os:pid, os:tid).

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>